+++ src/main/java/org/envyw/dadmarketplace/dto/searchKey/EquipmentDto.java
@@ -1,4 +1,4 @@
-package org.envyw.dadmarketplace.dto;
+package org.envyw.dadmarketplace.dto.searchKey;
 
 public record EquipmentDto(
         String name,
