+++ src/main/java/org/envyw/dadmarketplace/dto/searchKey/RarityDto.java
@@ -1,4 +1,4 @@
-package org.envyw.dadmarketplace.dto;
+package org.envyw.dadmarketplace.dto.searchKey;
 
 public record RarityDto(
         int id,
